Fix ROUND64/TRUNC64

+++ include/global.h
@@ -45,8 +45,8 @@
 
 #define ROUND(n, a) (((u32)(n) + (a)-1) & ~((a)-1))
 #define TRUNC(n, a) (((u32)(n)) & ~((a)-1))
-#define ROUND64(n, a) (((u64)(n) + (a)-1) & ~((a)-1))
-#define TRUNC64(n, a) (((u64)(n)) & ~((a)-1))
+#define ROUND64(n, a) (((u64)(n) + (a)-1) & ~(u64)((a)-1))
+#define TRUNC64(n, a) (((u64)(n)) & ~(u64)((a)-1))
 
 // Silence unused parameter warnings.
 // Necessary for debug matches.
